requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
amqp==1.4.9               # via -r requirements/base.txt, kombu
anyjson==0.3.3            # via -r requirements/base.txt, kombu
astroid==2.5.3            # via pylint
billiard==3.3.0.23        # via -r requirements/base.txt, celery
celery==3.1.26.post2      # via -c requirements/constraints.txt, -r requirements/base.txt
certifi==2020.12.5        # via -r requirements/base.txt, requests
chardet==4.0.0            # via -r requirements/base.txt, requests
coverage==5.5             # via -r requirements/test.in
ddt==1.4.2                # via -r requirements/test.in
django==2.2.20            # via -c requirements/constraints.txt, -r requirements/base.txt, jsonfield2
djangorestframework==3.9.4  # via -c requirements/constraints.txt, -r requirements/base.txt
edx-opaque-keys==2.1.0    # via -c requirements/constraints.txt, -r requirements/base.txt
idna==2.10                # via -r requirements/base.txt, requests
isort==4.3.21             # via -c requirements/constraints.txt, pylint
jsonfield2==3.0.3         # via -c requirements/constraints.txt, -r requirements/base.txt
kombu==3.0.37             # via -r requirements/base.txt, celery
lazy-object-proxy==1.6.0  # via astroid
mccabe==0.6.1             # via pylint
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in
oauthlib==3.0.1           # via -c requirements/constraints.txt, -r requirements/base.txt, requests-oauthlib
pbr==5.5.1                # via -r requirements/base.txt, stevedore
pycodestyle==2.7.0        # via -r requirements/test.in
pylint==2.7.4             # via -r requirements/test.in
pymongo==3.11.3           # via -r requirements/base.txt, edx-opaque-keys
pytz==2021.1              # via -r requirements/base.txt, celery, django
requests-oauthlib==1.3.0  # via -r requirements/base.txt
requests==2.25.1          # via -r requirements/base.txt, requests-oauthlib
six==1.15.0               # via -r requirements/base.txt, edx-opaque-keys, mock, stevedore
sqlparse==0.4.1           # via -r requirements/base.txt, django
stevedore==1.32.0         # via -c requirements/constraints.txt, -r requirements/base.txt, edx-opaque-keys
testfixtures==6.17.1      # via -r requirements/test.in
toml==0.10.2              # via pylint
urllib3==1.26.4           # via -r requirements/base.txt, requests
wrapt==1.12.1             # via astroid
xmltodict==0.12.0         # via -r requirements/base.txt
